chore: adopt conventional commits standard
The goal of adopting the conventional commits standard is to simplify
management of semantic versioning. The metadata provided by commit
messages can guide which releases are MAJOR, MINOR, and PATCH. As well
as give details needed to create CHANGELOGs in a rapid release cycle.
To further aid the process, this adopts the types and formatting
recommendations laid out by the Angular project.