fix(search): gracefully handle more directory search failure modalities (#825)
When directory search fails for e.g. bad directory search URL config,
fail with user-facing message and console log message, and stop the loading indicator.

Removes `$scope.wiscDirectoryError`

`wiscDirectoryErrorMessage` *is* the flag that there's an error, we don't need a separate `wiscDirectoryError` to have to keep in compatible state with the message.