fix: rm app directory search results count badge, temporarily (#843)
Mitigates bug in which the badge indicates more results than are visible to user,
hypothesis: because it's counting items on which user has BROWSE but not SUBSCRIBE
whereas actual results listing is only showing items on which user has both.

Intended as temporary workaround. Suppresses a misleading badge until can develop a deeper fix to reliably show the correct number in the badge.